fix(snap): remove incorrect usage of --target-arch for snapcraft (#41)
Snapcraft has ignored the target-arch flag, which works in very
limited circumstances, with destructive mode.  Snapcraft considers
it unsupported and will not be fixed in its current form.

Recently snapcraft regarded improper usage of this as a hard error
instead of silently ignoring, which broke users of app-builder. We
reverted that to simply warn, but this fix is to remove its usage
from app-builder.

If the user is targeting an arch that is not the host arch,
error out with a nice message.

However, it does require that --arch with `snap` must use the
scheme provided by go. If there are a set of supported architecture
options somewhere, I can update to ensure that they map to GOARCH.

Closes: #40

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>