Add support for darwin arm64 (#60)
Configures the build so that it can produce executables for darwin arm64, AKA "M1" Macs.

Because Go can't yet build universal binaries alone, and the Apple supplied tool to build universal binaries is unavailable
on non-Darwin platforms, we simply build and distribute versions of the executable for both architectures, and select the
appropriate one at runtime in Javascript. For backwards compatibility, an `app-builder` symlink to the amd64 executable is
maintained.