fix: handle the table of content resource type correctly (#148)
All ICNS resource type entries are made of a type and a length. The resource type is always a 4 byte string - if it's shorter than 4 bytes, it is filled with space (0x20) characters. This is also the case for the 'TOC ' resource type, so add the missing space character.

Modern ICNS tools often add a TOC\0x20 entry in the resource list, and that may not have been the case before. Add this to the fact that most linux-based builds are not based upon ICNS files and that should explain why this issue may not have been detected for 7 years (the fixed line was last changed in 2018).